requirements.txt
numpy==1.24.3
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
pillow==11.0.0
pyzmq==25.1.0
